+++ index.html
@@ -6,7 +6,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 
 	<!-- Font Awesome CSS -->
-	<link rel="stylesheet" href="node_modules/leaflet.locatecontrol/node_modules/font-awesome/css/font-awesome.min.css">
+	<link rel="stylesheet" href="node_modules/font-awesome/css/font-awesome.min.css">
 	
 	<!-- Leaflet CSS -->
 	<link rel="stylesheet" href="node_modules/leaflet/dist/leaflet.css">
