--- json/mobs.json
@@ -1,10 +1,0 @@
-[
-  {
-    "latitude": 48.262457,
-    "longitude": 11.669183
-  },
-  {
-    "latitude": 48.265882,
-    "longitude": 11.674327
-  }
-]
