+++ testMap.js
@@ -1,6 +1,6 @@
 
 var functions = require('./js/functions');
-var PokeMap = require('./js/opencyclemap');
+var opencyclemap = require('./js/opencyclemap');
 
 var sidebarHTML = '<div id="sidebar">\
 		<!--<div id="close"><a href="#" onclick="hideAdditionalInformation()">×</a></div>-->\
@@ -30,12 +30,10 @@ var elemDiv = document.createElement('div');
 elemDiv.innerHTML = sidebarHTML;
 document.body.appendChild(elemDiv);
 
-
-var poke = new PokeMap('map', [48.264673,11.671434], 17);
-poke.setUpMap(48.264673,11.671434);
+opencyclemap.setUpMap(48.264673,11.671434);
 //opencyclemap.setUpLocation(48.264673,11.671434);
 var from = new Date("2016-08-01T00:00:00.000Z");
 var to = new Date("2016-10-01T00:00:00.000Z");
-poke.loadPokemonData(poke.initializePokemonLayer, from, to);
+opencyclemap.loadPokemonData(opencyclemap.initializePokemonLayer, from, to);
 functions.initializeSlider();
